extracted preload test to a separate file
